Fixed a `ripgrep` problematic argument in TODOs
The argument "--ignore-files" was giving some errors (seems that dont exists, it is not even in the documentation).
If the pourpose was to ignore git files (from .gitignore), ripgrep ignores those files by [default](https://github.com/BurntSushi/ripgrep)
> By default, ripgrep will respect gitignore rules and automatically skip hidden files/directories and binary files

So tell me if you've considered any other variation of ripgrep. Or if that was a mistake, removing that arg is the solution I found after a while debugging.